+++ Utils/DeepL/TranslationResult.cs
@@ -1,0 +1,11 @@
+﻿using Newtonsoft.Json;
+
+namespace RevitTranslatorAddin.Utils.DeepL;
+public class TranslationResult
+{
+    [JsonProperty("translations")]
+    public Translation[] Translations
+    {
+        get; set;
+    }
+}
