+++ Utils/DeepL/TranslationUtils.cs
@@ -673,51 +673,3 @@ public class TranslationUtils
         }
     }
 }
-
-
-//TODO: Move to separate class files in "DeepL" directory
-
-/// <summary>
-/// Classes for handling responses from DeepL API.
-/// </summary>
-public class TranslationResult
-{
-    [JsonProperty("translations")]
-    public Translation[] Translations
-    {
-        get; set;
-    }
-}
-
-public class Translation
-{
-    [JsonProperty("detected_source_language")]
-    public string DetectedSourceLanguage
-    {
-        get; set;
-    }
-
-    [JsonProperty("text")]
-    public string Text
-    {
-        get; set;
-    }
-}
-
-/// <summary>
-/// Class that stores DeepL usage and quota.
-/// </summary>
-public class DeepLUsage
-{
-    [JsonProperty("character_count")]
-    public int CharacterCount
-    {
-        get; set;
-    }
-
-    [JsonProperty("character_limit")]
-    public int CharacterLimit
-    {
-        get; set;
-    }
-}
